Extract methods for JSON constructors to JsonConstructorUtils
