+++ h2/src/main/org/h2/expression/function/JsonConstructorFunction.java
@@ -15,10 +15,7 @@ import org.h2.expression.OperationN;
 import org.h2.expression.Subquery;
 import org.h2.expression.TypedValueExpression;
 import org.h2.message.DbException;
-import org.h2.util.json.JSONByteArrayTarget;
-import org.h2.util.json.JSONBytesSource;
-import org.h2.util.json.JSONStringTarget;
-import org.h2.util.json.JSONValidationTargetWithUniqueKeys;
+import org.h2.util.json.JsonConstructorUtils;
 import org.h2.value.TypeInfo;
 import org.h2.value.Value;
 import org.h2.value.ValueJson;
@@ -28,16 +25,6 @@ import org.h2.value.ValueNull;
  * JSON constructor function.
  */
 public final class JsonConstructorFunction extends OperationN implements ExpressionWithFlags, NamedExpression {
-
-    /**
-     * The ABSENT ON NULL flag.
-     */
-    public static final int JSON_ABSENT_ON_NULL = 1;
-
-    /**
-     * The WITH UNIQUE KEYS flag.
-     */
-    public static final int JSON_WITH_UNIQUE_KEYS = 2;
 
     private final boolean array;
 
@@ -80,62 +67,15 @@ public final class JsonConstructorFunction extends OperationN implements Express
             }
             Value value = args[i++].getValue(session);
             if (value == ValueNull.INSTANCE) {
-                if ((flags & JSON_ABSENT_ON_NULL) != 0) {
+                if ((flags & JsonConstructorUtils.JSON_ABSENT_ON_NULL) != 0) {
                     continue;
                 } else {
                     value = ValueJson.NULL;
                 }
             }
-            jsonObjectAppend(baos, name, value);
+            JsonConstructorUtils.jsonObjectAppend(baos, name, value);
         }
-        return jsonObjectFinish(baos, flags);
-    }
-
-    /**
-     * Appends a value to a JSON object in the specified string builder.
-     *
-     * @param baos
-     *            the output stream to append to
-     * @param key
-     *            the name of the property
-     * @param value
-     *            the value of the property
-     */
-    public static void jsonObjectAppend(ByteArrayOutputStream baos, String key, Value value) {
-        if (baos.size() > 1) {
-            baos.write(',');
-        }
-        JSONByteArrayTarget.encodeString(baos, key).write(':');
-        byte[] b = value.convertTo(TypeInfo.TYPE_JSON).getBytesNoCopy();
-        baos.write(b, 0, b.length);
-    }
-
-    /**
-     * Appends trailing closing brace to the specified string builder with a
-     * JSON object, validates it, and converts to a JSON value.
-     *
-     * @param baos
-     *            the output stream with the object
-     * @param flags
-     *            the flags ({@link #JSON_WITH_UNIQUE_KEYS})
-     * @return the JSON value
-     * @throws DbException
-     *             if {@link #JSON_WITH_UNIQUE_KEYS} is specified and keys are
-     *             not unique
-     */
-    public static Value jsonObjectFinish(ByteArrayOutputStream baos, int flags) {
-        baos.write('}');
-        byte[] result = baos.toByteArray();
-        if ((flags & JSON_WITH_UNIQUE_KEYS) != 0) {
-            try {
-                JSONBytesSource.parse(result, new JSONValidationTargetWithUniqueKeys());
-            } catch (RuntimeException ex) {
-                String s = JSONBytesSource.parse(result, new JSONStringTarget());
-                throw DbException.getInvalidValueException("JSON WITH UNIQUE KEYS",
-                        s.length() < 128 ? result : s.substring(0, 128) + "...");
-            }
-        }
-        return ValueJson.getInternal(result);
+        return JsonConstructorUtils.jsonObjectFinish(baos, flags);
     }
 
     private Value jsonArray(SessionLocal session, Expression[] args) {
@@ -148,7 +88,7 @@ public final class JsonConstructorFunction extends OperationN implements Express
                 if (arg0 instanceof Subquery) {
                     Subquery q = (Subquery) arg0;
                     for (Value value : q.getAllRows(session)) {
-                        jsonArrayAppend(baos, value, flags);
+                        JsonConstructorUtils.jsonArrayAppend(baos, value, flags);
                     }
                     break evaluate;
                 } else if (arg0 instanceof Format) {
@@ -157,43 +97,18 @@ public final class JsonConstructorFunction extends OperationN implements Express
                     if (arg0 instanceof Subquery) {
                         Subquery q = (Subquery) arg0;
                         for (Value value : q.getAllRows(session)) {
-                            jsonArrayAppend(baos, format.getValue(value), flags);
+                            JsonConstructorUtils.jsonArrayAppend(baos, format.getValue(value), flags);
                         }
                         break evaluate;
                     }
                 }
             }
             for (int i = 0; i < l;) {
-                jsonArrayAppend(baos, args[i++].getValue(session), flags);
+                JsonConstructorUtils.jsonArrayAppend(baos, args[i++].getValue(session), flags);
             }
         }
         baos.write(']');
         return ValueJson.getInternal(baos.toByteArray());
-    }
-
-    /**
-     * Appends a value to a JSON array in the specified output stream.
-     *
-     * @param baos
-     *            the output stream to append to
-     * @param value
-     *            the value
-     * @param flags
-     *            the flags ({@link #JSON_ABSENT_ON_NULL})
-     */
-    public static void jsonArrayAppend(ByteArrayOutputStream baos, Value value, int flags) {
-        if (value == ValueNull.INSTANCE) {
-            if ((flags & JSON_ABSENT_ON_NULL) != 0) {
-                return;
-            } else {
-                value = ValueJson.NULL;
-            }
-        }
-        if (baos.size() > 1) {
-            baos.write(',');
-        }
-        byte[] b = value.convertTo(TypeInfo.TYPE_JSON).getBytesNoCopy();
-        baos.write(b, 0, b.length);
     }
 
     @Override
@@ -235,14 +150,14 @@ public final class JsonConstructorFunction extends OperationN implements Express
      * @return the specified string builder
      */
     public static StringBuilder getJsonFunctionFlagsSQL(StringBuilder builder, int flags, boolean forArray) {
-        if ((flags & JSON_ABSENT_ON_NULL) != 0) {
+        if ((flags & JsonConstructorUtils.JSON_ABSENT_ON_NULL) != 0) {
             if (!forArray) {
                 builder.append(" ABSENT ON NULL");
             }
         } else if (forArray) {
             builder.append(" NULL ON NULL");
         }
-        if (!forArray && (flags & JSON_WITH_UNIQUE_KEYS) != 0) {
+        if (!forArray && (flags & JsonConstructorUtils.JSON_WITH_UNIQUE_KEYS) != 0) {
             builder.append(" WITH UNIQUE KEYS");
         }
         return builder;
